requirements.txt
django
psycopg2
django-leaflet
django-taggit
